requirements.txt
anyio==4.4.0
certifi==2024.2.2
dynaconf==3.2.5
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
python-telegram-bot==21.2
sniffio==1.3.1
typing_extensions==4.12.0
